add an example for kafka
